# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DealPilot Local API
  version: 1.0.0
  description: |
    DealPilot 本地优先架构的后端 API 契约。
    所有端点前缀 /api/v1，除 /health 外全部需要 Bearer Token 认证。
    Agent 只监听 127.0.0.1，禁止 0.0.0.0。
    列表响应使用游标分页（cursor + next_cursor）。
    写操作支持 Idempotency-Key 请求头实现幂等重试。
  contact:
    name: DealPilot
servers:
  - url: http://127.0.0.1:{port}
    description: 本地 Agent 回环地址
...
paths:
...
  /api/v1/reminders:
    get:
      tags: [reminders]
      summary: 提醒列表（待办，按排序规则）
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/ReminderStatus' }
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: sort_by
          in: query
          description: 排序字段（逾期/高风险/分级/到期时间）
          schema: { type: string, default: due_at }
      responses:
        '200':
          description: 提醒列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReminderListResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      tags: [reminders]
      summary: 新建提醒
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReminderCreate' }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Reminder' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/reminders/{id}:
    put:
      tags: [reminders]
      summary: 更新提醒状态（完成/稍后/忽略/已收到回复）
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReminderStatusUpdate' }
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Reminder' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/reminders/popup:
    get:
      tags: [reminders]
      summary: popup 前5条待办
      description: 按逾期/高风险/分级/到期时间排序，返回前5条待办。
      responses:
        '200':
          description: 待办列表
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Reminder' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/projects:
    get:
      tags: [projects]
      summary: 项目列表
      parameters:
        - name: customer_id
          in: query
          schema: { type: string }
        - name: stage
          in: query
          schema: { $ref: '#/components/schemas/ProjectStage' }
        - name: grade
          in: query
          schema: { $ref: '#/components/schemas/ProjectGrade' }
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: 项目列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectListResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      tags: [projects]
      summary: 新建项目
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectCreate' }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/projects/{id}:
    get:
      tags: [projects]
      summary: 项目详情
      description: 含风险和里程碑。
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 项目详情
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      tags: [projects]
      summary: 编辑项目
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectUpdate' }
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      tags: [projects]
      summary: 归档/关闭项目
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: 归档/关闭原因
              required: [reason]
      responses:
        '204':
          description: 归档/关闭成功
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/projects/{id}/stage:
    put:
      tags: [projects]
      summary: 切换项目阶段
      description: 用户手动切换，系统不自动推进阶段。
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stage: { $ref: '#/components/schemas/ProjectStage' }
              required: [stage]
      responses:
        '200':
          description: 切换成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/projects/{id}/risks:
    post:
      tags: [projects]
      summary: 添加风险
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RiskCreate' }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Risk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/risks/{id}:
    put:
      tags: [projects]
      summary: 更新风险状态
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RiskUpdate' }
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Risk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /api/v1/settings:
    get:
      tags: [settings]
      summary: 本地设置
      responses:
        '200':
          description: 设置对象
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Settings' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      tags: [settings]
      summary: 更新设置
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SettingsUpdate' }
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Settings' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string }
    CursorParam:
      name: cursor
      in: query
      description: 游标分页位置
      schema: { type: string }
    LimitParam:
      name: limit
      in: query
      description: 每页数量
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    IdempotencyKeyParam:
      name: Idempotency-Key
      in: header
      description: 幂等键，重复请求返回首次结果
      schema: { type: string }
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: 短期访问令牌，工作台来自 sessionStorage，插件来自 Native Messaging Bootstrap。
  responses:
    BadRequest:
      description: 请求参数或格式错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: 未认证或令牌无效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Origin 校验失败或权限不足
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: 资源冲突或状态非法
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    InternalError:
      description: 服务器内部错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: 错误码
        message:
          type: string
          description: 错误信息
        details:
          type: object
          description: 附加详情
        request_id:
          type: string
          description: 请求追踪 ID
      required: [code, message, request_id]
...
    ReminderType:
      type: string
      enum: [fixed_time, waiting_reply, paused]
    ReminderStatus:
      type: string
      enum: [pending, completed, snoozed, ignored, overdue, replied]
    ReminderPriority:
      type: string
      enum: [low, normal, high, urgent]
    Reminder:
      type: object
      properties:
        id: { type: string }
        customer_id: { type: string }
        project_id: { type: string, nullable: true }
        type: { $ref: '#/components/schemas/ReminderType' }
        status: { $ref: '#/components/schemas/ReminderStatus' }
        due_at: { type: string, format: date-time }
        priority: { $ref: '#/components/schemas/ReminderPriority' }
        last_notified_at: { type: string, nullable: true, format: date-time }
        snooze_until: { type: string, nullable: true, format: date-time }
        resolution: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, customer_id, type, status, due_at, priority, created_at, updated_at]
    ReminderCreate:
      type: object
      properties:
        customer_id: { type: string }
        project_id: { type: string }
        type: { $ref: '#/components/schemas/ReminderType' }
        due_at: { type: string, format: date-time }
        priority: { $ref: '#/components/schemas/ReminderPriority' }
      required: [customer_id, type, due_at]
    ReminderStatusUpdate:
      type: object
      properties:
        status: { $ref: '#/components/schemas/ReminderStatus' }
        snooze_until: { type: string, format: date-time }
        resolution: { type: string }
      required: [status]
    ReminderListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Reminder' }
        next_cursor: { type: string, nullable: true }
      required: [items]
    ProjectStage:
      type: string
      enum: [lead, qualified, proposal, negotiation, closed_won, closed_lost, archived]
    ProjectGrade:
      type: string
      enum: [S, A, B, C]
    Project:
      type: object
      properties:
        id: { type: string }
        customer_id: { type: string }
        name: { type: string }
        currency: { type: string }
        amount: { type: number, nullable: true }
        probability: { type: integer, minimum: 0, maximum: 100, nullable: true }
        expected_close_date: { type: string, nullable: true, format: date }
        stage: { $ref: '#/components/schemas/ProjectStage' }
        grade: { $ref: '#/components/schemas/ProjectGrade' }
        closed_reason: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, customer_id, name, stage, grade, created_at, updated_at]
    ProjectCreate:
      type: object
      properties:
        customer_id: { type: string }
        name: { type: string }
        currency: { type: string }
        amount: { type: number }
        probability: { type: integer }
        expected_close_date: { type: string, format: date }
        stage: { $ref: '#/components/schemas/ProjectStage' }
        grade: { $ref: '#/components/schemas/ProjectGrade' }
      required: [customer_id, name, stage, grade]
    ProjectUpdate:
      type: object
      properties:
        name: { type: string }
        currency: { type: string }
        amount: { type: number }
        probability: { type: integer }
        expected_close_date: { type: string, format: date }
        stage: { $ref: '#/components/schemas/ProjectStage' }
        grade: { $ref: '#/components/schemas/ProjectGrade' }
        closed_reason: { type: string }
    ProjectDetail:
      type: object
      allOf:
        - $ref: '#/components/schemas/Project'
        - type: object
          properties:
            risks:
              type: array
              items: { $ref: '#/components/schemas/Risk' }
            milestones:
              type: array
              items: { $ref: '#/components/schemas/Milestone' }
            pending_reminders:
              type: array
              items: { $ref: '#/components/schemas/Reminder' }
    ProjectListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Project' }
        next_cursor: { type: string, nullable: true }
      required: [items]
    RiskSeverity:
      type: string
      enum: [low, medium, high, critical]
    RiskStatus:
      type: string
      enum: [open, handling, resolved, ignored]
    Risk:
      type: object
      properties:
        id: { type: string }
        project_id: { type: string }
        description: { type: string }
        severity: { $ref: '#/components/schemas/RiskSeverity' }
        status: { $ref: '#/components/schemas/RiskStatus' }
        handled_at: { type: string, nullable: true, format: date-time }
        created_at: { type: string, format: date-time }
      required: [id, project_id, description, severity, status, created_at]
    RiskCreate:
      type: object
      properties:
        description: { type: string }
        severity: { $ref: '#/components/schemas/RiskSeverity' }
      required: [description, severity]
    RiskUpdate:
      type: object
      properties:
        status: { $ref: '#/components/schemas/RiskStatus' }
        severity: { $ref: '#/components/schemas/RiskSeverity' }
        handled_at: { type: string, format: date-time }
    Milestone:
      type: object
      properties:
        id: { type: string }
        project_id: { type: string }
        name: { type: string }
        date: { type: string, format: date }
        completed: { type: boolean }
        created_at: { type: string, format: date-time }
      required: [id, project_id, name, date, completed, created_at]
...
    Settings:
      type: object
      properties:
        id: { type: integer, example: 1 }
        last_backup_at: { type: string, nullable: true, format: date-time }
        auto_start: { type: boolean }
        minimize_to_tray: { type: boolean }
        backup_reminder_days: { type: integer, nullable: true }
        locale: { type: string }
        theme: { type: string }
      required: [id]
    SettingsUpdate:
      type: object
      properties:
        last_backup_at: { type: string, format: date-time }
        auto_start: { type: boolean }
        minimize_to_tray: { type: boolean }
        backup_reminder_days: { type: integer }
        locale: { type: string }
        theme: { type: string }
...
security:
  - BearerAuth: []